manually do a typecast to support non full featured c++ 11

+++ VCMTPv3/sender/TcpSend.cpp
@@ -75,7 +75,7 @@ TcpSend::TcpSend(string tcpAddr, unsigned short tcpPort)
     if(::bind(sockfd, (struct sockaddr *) &servAddr, sizeof(servAddr)) < 0)
         throw std::system_error(errno, std::system_category(),
                 "TcpSend::TcpSend(): Couldn't bind \"" + tcpAddr + ":" +
-                std::to_string(tcpPort) + "\"");
+                std::to_string(static_cast<long long>(tcpPort)) + "\"");
     /** listen() returns right away, it's non-blocking */
     listen(sockfd, MAX_CONNECTION);
 }
